campusassist/backend/requirements.txt
flask
flask-cors
twilio
python-dotenv
google-generativeai
